requirements.txt
pandas
numpy
statsmodels
scikit-learn
matplotlib
psycopg2-binary
python-dotenv
streamlit
plotly
pytz
joblib
schedule
